feat: Correct click-to-schema logic
problem: It is necessary to unify the behavior of the GUI so that the pressed
button in the tool bar is related to each other and at the same time
to distinguish the first and second click of the mouse.

solution: With the first click, while holding down the button in the tool bar,
we activate the function, mark the starting point, with the second click we
mark the end point.

next step: Now we need to design a class that will hold the data and finish
rendering of the starting point, the drawn line and the ending point.