refactor: _ManuallyBufferedPaintDC()
problem: wxPython uses different DC in different situation. In case of
wx.EVT_PAINT, we should use wx.AutoBufferedPaintDC(). But in case of
wx.EVT_MOTION, we are using manualy buffered memory.

As we progressing with editor, we want to draw also from wx.EVT_LEFT_DOWN
which means we need to separate manual memory buffering for DC to
private method.