refactor(Painter): cursor_position --> cursor_tile
problem: To avoid constant recalculation between x, y and tile,
the Painter class should receive the cursor position in the tile. Due to
the fact that we instantiate the Painter class again each time we draw,
the cursor position should be passed to the constructor (we never call
the draw method multiple times on the same instance of the Painter
class).